requirements.txt
Flask==3.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
langdetect==1.0.9
python-docx==1.0.1
Werkzeug==3.0.1
Flask-Session==0.5.0
Flask-WTF==1.2.1
gunicorn==21.2.0
